dev-requirements.txt
nose
pypiwin32
Pillow==2.7.0
coverage
codecov
pyscreenshot